Add argument for arbitrary RemoteProvider kwargs
The --default-remote-provider and --default-remote-prefix arguments are
insufficient for using SFTP as the default remote provider. For example,
there is no way to specify a username and password with those arguments.

This commit adds a --default-remote-kwargs argument that allows
arbitrary keyword-arguments to be passed from the command-line to the
remote provider constructor. In the case of the SFTP provider, these
arguments could include `username` or `password`.